% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lua51-package.R
\docType{package}
\name{lua51-package}
\alias{lua51}
\alias{lua51-package}
\title{lua51: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
